move code into Game component
