+++ src/app/page.tsx
@@ -1,35 +1,11 @@
 "use client"
 
-import { useState } from "react";
-import HighlightText from "./HighlightText";
+import Game from "./Game";
 
 export default function Home() {
-
-  const text = "one two three four five six seven";
-  const words = text.split(/(?<=\s)/); //split and keep whitespace character
-
-  const [input, setInput] = useState('');
-  const [wordIndex, setWordIndex] = useState(0);
-
-  const [completedText, setCompletedText] = useState('');
-
-  const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    let currentWord = words[wordIndex];
-    let val = e.target.value;
-
-    if(val === currentWord) {
-      setCompletedText(userText => userText + currentWord)
-      setWordIndex(wordIndex => wordIndex + 1)
-      setInput('');
-    } else {
-      setInput(e.target.value);
-    }
-  } 
-
   return (
     <main>
-      <HighlightText value={text} compareString={completedText + input}/>
-      <input autoFocus value={input} onChange={onChange}/>
+      <Game/>
     </main>
   );
 }
